requirements.txt
# -----------------------------
# Core Web Framework
# -----------------------------
Flask==3.0.3

# -----------------------------
# Data Processing
# -----------------------------
pandas==2.2.1
numpy==1.26.4

# -----------------------------
# NLP & AI Models
# -----------------------------
spacy==3.7.2
transformers==4.39.3
sentencepiece==0.1.99
accelerate==0.27.2

# LangChain (requires Pydantic v2)
langchain==0.3.14
langchain-core==0.3.30
langchain-community==0.3.14


# -----------------------------
# ML Utilities
# -----------------------------
scikit-learn==1.4.2
joblib==1.3.2

# -----------------------------
# Visualization
# -----------------------------
matplotlib==3.8.4
seaborn==0.13.2

# -----------------------------
# Video & Animation
# -----------------------------
moviepy==1.0.3
Pillow==10.2.0
imageio==2.34.0
imageio-ffmpeg==0.5.1

# -----------------------------
# Audio Generation (Optional)
# -----------------------------
gTTS==2.3.2

# -----------------------------
# Backend API Tools
# -----------------------------
requests==2.31.0

# -----------------------------
# Pydantic v2 Ecosystem
# -----------------------------
pydantic==2.7.4
pydantic-core==2.18.4

# -----------------------------
# Utilities
# -----------------------------
python-dotenv==1.0.1
tqdm==4.66.1
